+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2021-12-08
+### Added
+-   Fetch photographer data and filter them according to the current photographer "ID".
+-   Display number of likes of the first current photographer media (Work in progress. Needs to refactor with "Factory pattern).
+
 ## [0.1.2] - 2021-12-08
 ### Added
 -   Display photographer profile data on photographer page (Two methods available).